plugins/admin/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qodalis-cli-server-plugin-admin"
version = "1.0.0b1"
description = "Admin dashboard plugin for Qodalis CLI servers"
readme = { text = "", content-type = "text/plain" }
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nicolae Lupei", email = "nicolae.lupei@qodalis.com" },
]
keywords = ["cli", "qodalis", "admin", "dashboard", "plugin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "PyJWT>=2.0.0",
    "fastapi>=0.100.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/qodalis-solutions/cli-server-python"
Repository = "https://github.com/qodalis-solutions/cli-server-python"
Issues = "https://github.com/qodalis-solutions/cli-server-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["qodalis_cli_admin"]

[tool.hatch.build.targets.sdist]
include = ["qodalis_cli_admin/**"]
